+++ projects/swimlane/docspa-core/src/lib/modules/markdown-elements/embed-file.ts
@@ -1,7 +1,7 @@
 import {
   Component, ViewEncapsulation,
   OnInit, OnChanges, HostBinding, HostListener,
-  Input, Output, EventEmitter,
+  Input,
   ElementRef, Renderer2,
   SimpleChanges
 } from '@angular/core';
@@ -10,9 +10,14 @@ import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
 import { MarkdownService } from '../markdown/markdown.service';
 import { LocationService } from '../../services/location.service';
 import { RouterService } from '../../services/router.service';
+import { FetchService } from '../../services/fetch.service';
+import { HooksService } from '../../services/hooks.service';
 import { splitHash } from '../../utils';
 
-import * as vfile from 'vfile';
+import { VFile } from '../../../vendor';
+import VFILE from 'vfile';
+
+const codefilesTypes = ['js', 'json'];
 
 @Component({
   selector: 'docspa-md-include', // tslint:disable-line
@@ -26,10 +31,15 @@ export class EmbedMarkdownComponent implements OnInit, OnChanges {
   path = '';
 
   @Input()
-  plugins = false;
+  isPageContent = false;
 
   @Input()
-  safe = false;
+  set safe(val: boolean) {
+    this._safe = true;
+  }
+  get safe() {
+    return this._safe === null ? this.isPageContent : this._safe;
+  }
 
   @Input()
   scrollTo: string;
@@ -43,7 +53,8 @@ export class EmbedMarkdownComponent implements OnInit, OnChanges {
   @Input()
   codeblock: string = null;
 
-  @Output() done: EventEmitter<vfile.VFile> = new EventEmitter();
+  @Input()
+  attr: string = null;
 
   @HostBinding('innerHTML')
   html: string | SafeHtml;
@@ -51,13 +62,17 @@ export class EmbedMarkdownComponent implements OnInit, OnChanges {
   @HostListener('click', ['$event'])
   onClickBtn = this.routerService.clickHandler;
 
+  private _safe: boolean = null;
+
   constructor(
     private markdownService: MarkdownService,
     private routerService: RouterService,
     private sanitizer: DomSanitizer,
     private elm: ElementRef,
     private renderer: Renderer2,
-    private locationService: LocationService
+    private locationService: LocationService,
+    private fetchService: FetchService,
+    private hooks: HooksService,
   ) {
   }
 
@@ -75,29 +90,57 @@ export class EmbedMarkdownComponent implements OnInit, OnChanges {
   }
 
   ngOnInit() {
-    this.load();
+    this.load().then(() => {
+      this.hooks.doneEach.tap('update-links', () => {
+        this.markActiveLinks();
+      });
+    });
   }
 
-  private load() {
-    if (!this.codeblock) {
-      return this.markdownService.getMd(this.path, this.plugins)
-        .subscribe(_vfile => {
-          setTimeout(() => {
-            this.markActiveLinks();
-            this.doScroll();
-            this.done.emit(_vfile);
-          }, 30);
-          this.html = this.safe ? this.sanitizer.bypassSecurityTrustHtml(_vfile.contents as string) : _vfile.contents;
-        });
+  private async load(): Promise<string | SafeHtml> {
+    const _vfile = await this.fetch();
+
+    const { notFound } = _vfile.data.docspa;
+
+    const ext = _vfile.extname.replace('.', '');
+
+    let codeblock = this.codeblock;
+    const bypassSecurity = this.safe && !notFound;
+
+    if (!codeblock && codefilesTypes.includes(ext)) {
+      codeblock = ext;
     }
 
-    const vf: any = this.locationService.pageToFile(this.path);
-    return fetch(vf.data.docspa.url).then(res => res.text()).then(async contents => {
-      vf.contents = `~~~${this.codeblock}\n${contents}\n~~~`;
-      await this.markdownService.processor.process(vf);
-      await this.markdownService.processMd(true, vf);
-      this.html = this.safe ? this.sanitizer.bypassSecurityTrustHtml(vf.contents) : vf.contents;
-    });
+    if (codeblock) {
+      codeblock = codeblock + (this.attr ? ` ${this.attr}` : '');
+      _vfile.contents = `~~~${codeblock}\n${_vfile.contents}\n~~~`;
+    }
+
+    await this.markdownService.process(_vfile);
+    setTimeout(() => {
+      this.markActiveLinks();
+      this.doScroll();
+      this.hooks.doneEach.call(_vfile);
+    }, 30);
+
+    this.html = bypassSecurity ? this.sanitizer.bypassSecurityTrustHtml(_vfile.contents as string) : _vfile.contents;
+    return this.html;
+  }
+
+  private async fetch() {
+    let _vfile: VFile;
+
+    if (!this.path) {
+      _vfile = VFILE('') as VFile;
+      _vfile.data.docspa.notFound = true;
+    } else {
+      _vfile = this.locationService.pageToFile(this.path) as VFile;
+      const { contents, notFound } = await this.fetchService.get(_vfile.data.docspa.url).toPromise();
+      _vfile.data.docspa.notFound = notFound;
+      _vfile.data.docspa.isPageContent = this.isPageContent;
+      _vfile.contents = (!notFound || this.isPageContent) ? contents : `!> *File not found*\n!> ${this.path}`;
+    }
+    return _vfile;
   }
 
   // Scroll to current anchor
@@ -129,7 +172,6 @@ export class EmbedMarkdownComponent implements OnInit, OnChanges {
   private markActiveLinks() {
     if (this.activeLink) {
       const aObj = this.elm.nativeElement.getElementsByTagName('a');
-
       const activeLinks = [];
       for (let i = 0; i < aObj.length; i++) {
         const a = aObj[i];