+++ projects/swimlane/docspa-core/src/lib/services/hooks.service.ts
@@ -6,10 +6,16 @@ import { AsyncSeriesWaterfallHook, SyncHook } from 'tapable';
 })
 export class HooksService {
   init = new SyncHook(); // Called when the script starts running, only trigger once, no arguments
-  beforeEach = new AsyncSeriesWaterfallHook(['vfile']);  // Invoked each time before parsing the Markdown file.
-  afterEach = new AsyncSeriesWaterfallHook(['vfile']); // Invoked each time after the Markdown file is parsed.
-  doneEach = new SyncHook(['vfile']); // Invoked each time after the data is fully loaded
+
+  // Invoked while processing markdown
+  beforeEach = new AsyncSeriesWaterfallHook(['vfile']);  // MD transform invoked each time before parsing the Markdown file.
+  afterEach = new AsyncSeriesWaterfallHook(['vfile']); // HTML transform invoked each time after the Markdown file is parsed.
+
+  // Invoked when dom is updated
+  doneEach = new SyncHook(['vfile']); // Invoked each time new main content is loaded (DOM changes)
+
+  // Life cycle
   mounted = new SyncHook(); // Called after initial completion.  (ngOnInit)
   ready = new SyncHook(); // Called after initial completion.  (ngAfterViewInit)
-  destroy = new SyncHook(); // Called after initial completion.  (ngOnDestroy)
+  destroy = new SyncHook(); // Called after when page is destroyed.  (ngOnDestroy)
 }